[BREAK] shipper: remove Variant model
This is a BREAKING change. Delete your current developmental
database as the existing migrations have been deleted.

The Variant model is being removed because it has been decided
to check GApps inclusion in the Build model. However, since the
relation chain breaks with this change, the pre-existing
migrations have been deleted. We haven't put anything into
production yet so this should be fine.