Release 1.2.4
There are several major configuration changes in this release.
This can be considered a breaking change - if you update blindly,
your installation will be broken!

Changes since 1.2.3:
 - The production and development Dockerfiles and configurations
   have been unified. Please refer to the wiki for more information.
   You will need to edit your configuration file names, but other
   than that no further changes are necessary.
 - shippy is no longer included with the main repository.
 - nginx is now directly downloaded and used without building.

In the future the shipper image will be used in the Docker Compose
configuration, negating the need to build the images locally. This
should help speed up deployment and make it easier for admins to
provision shipper!