Release 1.8.3
READ THIS AND THE [UPGRADE WIKI PAGE](https://github.com/ericswpark/shipper/wiki/Upgrading) VERY CAREFULLY!

shipper is moving to a custom user model, but limitations in Django make this a very difficult move. Therefore, this is an in-between release in order to keep databases intact. You must follow instructions precisely to avoid the loss of data.

If you are on version 1.8.2 or lower, you MUST upgrade to this version AND run a migration command BEFORE upgrading to the next release (1.9.0). If you skip this version and attempt to upgrade to the next release directly, the migration will fail and your database may get corrupted.

Again, if you are on version 1.8.2 or lower:

1. Upgrade to 1.8.3 (this release)
2. RUN A DATABASE MIGRATION (extremely important)
3. Then upgrade to the next release (1.9.0).