License shipper under GPLv3
shipper is now licensed under GPLv3, which means it is now a complete
FOSS software/repository! The reason I chose GPLv3 instead of another
license like MIT is because I wish for future improvements to be
submitted back upstream so it can help everyone using this repo.

Since I'm the only one working on this repository re-licensing the
project might be possible, but we'll cross that bridge once we get 
to it and once we feel the need to. (This is NOT an indication that
the project is dual-licensed -- if the project's license changes
there will be a new commit and this file's (LICENSE) contents
will change.