config: settings: add numbers and underscores to default filename regex
The default filename regex pattern only allows alphabets in the
codename, when codenames can contain numeric values and underscores,
like in the case of `Z01R` or `laurel_sprout`. While this option
is customizable by the end-user (admin), set defaults to this sensible
option so future admins don't hit on this snag when they encounter
a device codename with underscores and numbers.