shipper: models: check list length when querying build for variant
Because we were not checking the list length when fetching the
first item in the list of the latest builds for a given variant,
we were running into a IndexError because the length was less than 0,
and causing a 500 server error. The fix checks the list length and
raises a DoesNotExist error if the build list is empty.