Switch license to AGPLv3
Since shipper is server-side software, it makes no sense to provide it under GPL, 
since end users do not technically get a "copy" of the software. To fix this, change
the licensing of shipper to AGPLv3 instead.

Since the sole contributor to this repository is me (@ericswpark), I am authorizing
the re-licensing of this repository to AGPLv3.