feat: new task to clean up failed mirror build task results
This is a dirty, hacky workaround, because Celery stuff in the
django-celery-results repo do not support async/await, so some task
results are indefinitely stuck with PROGRESS, which is not what we want.
So this task manually checks each task result against the configured
time limit, and manually marks each task as FAILED if they have (gone
past the limit). Get rid of this workaround as soon as possible. (Burn
it.)