refactor(server): rewrite device and build model test setup
This significantly reduces the number of "magic numbers" sprinkled
everywhere, which makes adding new devices and builds easier since less
stuff is hardcoded. Also trims down a ton of repetitive code, which can
be a source of bugs in of itself.

The only minor problem I see with this approach is that the setup for
the devices and the builds isn't perfect. If a build variant was
previously defined, then you need to remember to use the `+=` operator
instead of the `=` one, which will overwrite the value and cause a bug.
Personally, I don't think this is a big issue, because you will probably
run the test suite again after making changes and if you used the wrong
operator then a ton of tests will fail. Still, probably a design that
can be improved later.