refactor(server): cleaner bundled JS generation
Add a .bundled extension to all generated JS files, so we don't have to
keep editing the .gitignore every time we add a new React part to the
project. In the future we should probably look into cleaning this up
even more, as I don't think this is the optimal solution.