fix(server): delete invalid chunked uploads
If the chunked upload is already marked as completed, or if it has
expired, then it can get stuck in a "limbo" state where it is
simultaneously processed and not processed. This state prevents the
maintainers from retrying the failed upload. Therefore, if any errors
are emitted in the finalization process, deem the chunked upload failed
and delete it.