feat(server): one-click statistics and mirror download URL (#436)
This replaces the distributed URL being sent out via the API. The new
endpoint creates the statistics item on download, as well as redirecting
users to the final URL that contains the build file.

The main website download buttons still allow the user to choose the
mirrors, although we should probably experiment with replacing it with
this endpoint instead. That is for another day/feature request.

Closes #436