fix(docker): use proper nginx config for layered reverse proxies (#562)
In case the setup utilizes multiple reverse proxies, the X-Forwarded-*
headers will be mangled with the previous configuration. However, to
trust the X-Forwarded-* headers, we need to trust that there will be an
edge reverse proxy serving requests, and that the reverse proxy is
correctly clearing and setting the X-Forwarded-* headers before
propagating them down the stack.

With the configuration change, make a note about the potential security
vulnerability with a possible unsupported setup, and the importance of
passing verified X-Forwarded-* headers in the documentation.

Fixes #562